added demonstration for handlers
